Students data
This is a code where you put student data, such as age, shift, and names, and then it gives you information and comments about specific things. 
Those things are:

Arithmetic mean age
The number of students that are younger than fifteen years old and study in the morning
The oldest person's name
And the number of students that study in the evening

It's all in portuguese, but i promisse it's accurate.